# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Cloudpass
  description: User management service
  contact:
    email: msuret@dhatim.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:10010
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
produces:
  - application/json
paths:
...
  /tenants/{id}/customData:
    x-swagger-router-controller: tenantController
    get:
      summary: retrieve custom data
      description: retrieve custom data
      operationId: getCustomData
      tags:
        - tenants
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update tenant custom data
      description: update custom data of the tenant with the given ID using the provided values
      operationId: updateCustomData
      tags:
        - tenants
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete custom data
      description: delete custom data
      operationId: deleteCustomData
      tags:
        - tenants
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /tenants/{id}/customData/{fieldName}:
    x-swagger-router-controller: tenantController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - tenants
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data field deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /tenants/{id}/{collection}:
    x-swagger-router-controller: tenantController
    get:
      summary: retrieve a collection sub-resource from the tenant
      description: retrieve a collection sub-resource from the tenant
      operationId: getCollection
      tags:
        - tenants
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - applications
            - directories
            - groups
            - accounts
            - organizations
            - idSites
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /applications/{id}/customData:
    x-swagger-router-controller: applicationController
    get:
      summary: retrieve custom data
      description: retrieve custom data
      operationId: getCustomData
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update custom data
      description: update custom data
      operationId: updateCustomData
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete custom data
      description: delete custom data
      operationId: deleteCustomData
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /applications/{id}/customData/{fieldName}:
    x-swagger-router-controller: applicationController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data field deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /applications/{id}/{collection}:
    x-swagger-router-controller: applicationController
    get:
      summary: retrieve a collection sub-resource from the application
      description: retrieve a collection sub-resource from the application
      operationId: getCollection
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - accountStoreMappings
            - directories
            - organizations
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /directories/{id}/customData:
    x-swagger-router-controller: directoryController
    get:
      summary: retrieve custom data
      description: retrieve custom data
      operationId: getCustomData
      tags:
        - directories
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update directory custom data
      description: update directory custom data
      operationId: updateCustomData
      tags:
        - directories
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete directory custom data
      description: delete directory custom data
      operationId: deleteCustomData
      tags:
        - directories
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /directories/{id}/customData/{fieldName}:
    x-swagger-router-controller: directoryController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - directories
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data field deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"    
...
  /directories/{id}/{collection}:
    x-swagger-router-controller: directoryController
    get:
      summary: retrieve a collection sub-resource from the directory
      description: retrieve a collection sub-resource from the directory
      operationId: getCollection
      tags:
        - directories
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - applicationMappings
            - applications
            - organizations
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /groups/{id}/customData:
    x-swagger-router-controller: groupController
    get:
      summary: retrieve group custom data
      description: retrieve group custom data
      operationId: getCustomData
      tags:
        - groups
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update group custom data
      description: update group custom data
      operationId: updateCustomData
      tags:
        - groups
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete custom data
      description: delete custom data
      operationId: deleteCustomData
      tags:
        - groups
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /groups/{id}/customData/{fieldName}:
    x-swagger-router-controller: groupController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - groups
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"   
  /groups/{id}/{collection}:
    x-swagger-router-controller: groupController
    get:
      summary: retrieve a collection sub-resource from the group
      description: retrieve a collection sub-resource from the group
      operationId: getCollection
      tags:
        - groups
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - accountMemberships
            - applicationMappings
            - applications
            - accounts
            - organizations
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /accounts/{id}/customData:
    x-swagger-router-controller: accountController
    get:
      summary: retrieve account custom data
      description: retrieve account custom data
      operationId: getCustomData
      tags:
        - accounts
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update account custom data
      description: update account custom data
      operationId: updateCustomData
      tags:
        - accounts
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete custom data
      description: delete custom data
      operationId: deleteCustomData
      tags:
        - accounts
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /accounts/{id}/customData/{fieldName}:
    x-swagger-router-controller: accountController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - accounts
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data field deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"   
  /accounts/{id}/{collection}:
    x-swagger-router-controller: accountController
    get:
      summary: retrieve a collection sub-resource from the account
      description: retrieve a collection sub-resource from the account
      operationId: getCollection
      tags:
        - accounts
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - groupMemberships
            - groups
            - applications
            - organizations
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /passwordPolicies/{id}/{collection}:
    x-swagger-router-controller: passwordPolicyController
    get:
      summary: retrieve a collection sub-resource from the password policy
      description: retrieve a collection sub-resource from the password policy
      operationId: getCollection
      tags:
        - password policies
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - resetEmailTemplates
            - resetSuccessEmailTemplates
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /accountCreationPolicies/{id}/{collection}:
    x-swagger-router-controller: accountCreationPolicyController
    get:
      summary: retrieve a collection sub-resource from the policy
      description: retrieve a collection sub-resource from the policy
      operationId: getCollection
      tags:
        - account creation policies
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - verificationEmailTemplates
            - verificationSuccessEmailTemplates
            - welcomeEmailTemplates
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
  /organizations/{id}/customData:
    x-swagger-router-controller: organizationController
    get:
      summary: retrieve custom data
      description: retrieve custom data
      operationId: getCustomData
      tags:
        - organizations
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "200":
          description: found
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    post:
      summary: update custom data
      description: update custom data
      operationId: updateCustomData
      tags:
        - organizations
      parameters:
        - $ref: "#/parameters/Id"
        - name: newCustomData
          description: New custom data
          in: body
          required: true
          schema:
            type: object
      responses:
        "200":
          description: updated
          schema:
            $ref: "#/definitions/Resource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
    delete:
      summary: delete custom data
      description: delete custom data
      operationId: deleteCustomData
      tags:
        - applications
      parameters:
        - $ref: "#/parameters/Id"
      responses:
        "204":
          description: custom data deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /organizations/{id}/customData/{fieldName}:
    x-swagger-router-controller: organizationController
    delete:
      summary: delete custom data field
      description: delete custom data field
      operationId: deleteCustomDataField
      tags:
        - organizations
      parameters:
        - $ref: "#/parameters/Id"
        - name: fieldName
          in: path
          description: field to delete
          required: true
          type: string
      responses:
        "204":
          description: custom data field deleted
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
  /organizations/{id}/{collection}:
    x-swagger-router-controller: organizationController
    get:
      summary: retrieve a collection sub-resource from the organization
      description: retrieve a collection sub-resource from the organization
      operationId: getCollection
      tags:
        - organizations
      parameters:
        - $ref: "#/parameters/Id"
        - name: collection
          in: path
          description: the collection to retrieve
          required: true
          type: string
          enum:
            - accountStoreMappings
            - directories
            - applications
        - $ref: "#/parameters/Offset"
        - $ref: "#/parameters/Limit"
        - $ref: "#/parameters/OrderBy"
        - $ref: "#/parameters/Expand"
      responses:
        "200":
          description: ok
          schema:
            $ref: "#/definitions/CollectionResource"
        "default":
          description: error
          schema:
            $ref: "#/definitions/Error"
...
definitions:
  Resource:
    type: object
    required:
      - href
      - createdAt
      - modifiedAt
    properties:
      href:
        type: string
      createdAt:
        type: string
        format: date-time
      modifiedAt:
        type: string
        format: date-time
  CollectionResource:
    type: object
    required:
      - href
      - offset
      - limit
      - size
      - items
    properties:
      href:
        type: string
      offset:
        type: integer
      limit:
        type: integer
      size:
        type: integer
      items:
        type: array
        items:
          $ref: "#/definitions/Resource"
...
  Error:
    type: object
    required:
      - status
      - code
      - message
      - developerMessage
      - moreInfo
    properties:
      status:
        type: integer
      code:
        type: integer
      message:
        type: string
      developerMessage:
        type: string
      moreInfo:
        type: string
...
parameters:
  Id:
    name: id
    in: path
    description: Resource ID
    required: true
    type: string
    format: uuid
...
  Expand:
    name: expand
    in: query
    description: the links to expand
    required: false
    type: string
  Offset:
    name: offset
    in: query
    description: resource collection offset
    required: false
    type: integer
    default: 0
  Limit:
    name: limit
    in: query
    description: resource collection limit
    required: false
    type: integer
    default: 25
  OrderBy:
    name: orderBy
    in: query
    description: resource collection ordering
    required: false
    type: array
    collectionFormat: csv
    items:
      type: string
...